refactor(terminal): split the manifest model out of its writer
An audit rejected the work because terminal-manifest.ts had grown to 308
lines against this plan's 250-line ceiling for new files - it was exactly
250 when it landed, then the durable re-adoption fix added +53.

The manifest DATA MODEL now lives in terminal-manifest-model.ts: the
version and debounce constants, the session type, the runtime-kind and
durability-class aliases, and the seven persisted interfaces. The store
factory and TerminalManifestWriter stay put and re-export every moved
symbol, so no importer's import list moved.

Both files are now under the ceiling (248 and 95), and the split is by
responsibility rather than by line count: the model is what the manifest
IS, the writer is how it changes.
