+++ packages/coding-agent/src/core/extensions/builtin/terminal/terminal-manifest.ts
@@ -1,104 +1,45 @@
 /**
- * Per-session terminal manifest: the durable sidecar record of live monitors and background
- * sessions that `restore.ts` reads back after a restart. The writer persists ONLY lifecycle
- * transitions plus debounced checkpoints — never per-line output and never a runtime handle.
+ * The manifest write side: `createTerminalManifestStore` plus `TerminalManifestWriter`,
+ * persisting ONLY lifecycle transitions and debounced checkpoints of the record shaped in
+ * `terminal-manifest-model.ts` — never per-line output and never a runtime handle.
  */
 
 import { join } from "node:path";
-import type { SessionManager } from "../../../session-manager.ts";
 import { createSidecarStore, type SidecarStore } from "../../../session-sidecar-store.ts";
 import type { MonitorSnapshotEntry } from "./monitor-registry.ts";
 import { parseTerminalManifest } from "./restore.ts";
 import { DURABLE_MONITOR_EXPIRY_MS } from "./shared.ts";
+import {
+	type ManifestBackgroundSession,
+	type ManifestMonitor,
+	type MonitorRegistration,
+	TERMINAL_MANIFEST_CHECKPOINT_DEBOUNCE_MS,
+	TERMINAL_MANIFEST_VERSION,
+	type TerminalManifest,
+	type TerminalManifestCheckpoint,
+	type TerminalManifestSession,
+} from "./terminal-manifest-model.ts";
 
-export const TERMINAL_MANIFEST_VERSION = 1;
-/** Minimum debounce for checkpoint writes; a burst inside this window collapses to one write. */
-export const TERMINAL_MANIFEST_CHECKPOINT_DEBOUNCE_MS = 30_000;
-/**
- * Re-exported so manifest consumers read one durability deadline. The value lives in
- * `shared.ts` with the rest of the terminal caps; there is exactly one definition of it.
- */
-export { DURABLE_MONITOR_EXPIRY_MS };
-export type TerminalManifestSession = Pick<SessionManager, "getSessionDir" | "getSessionId">;
-export type MonitorRuntimeKind = "command" | "file";
-export type MonitorDurabilityClass = "ephemeral" | "restartable-command" | "checkpointed-file";
-/**
- * File-monitor checkpoint persisted by the writer: the registry's live identity tuple. `digest`
- * is required — without it a same-size, same-mtime rewrite is undetectable across a restart.
- */
-export interface TerminalManifestCheckpoint {
-	readonly dev: number;
-	readonly ino: number;
-	readonly size: number;
-	readonly mtimeMs: number;
-	readonly digest: string;
-	readonly present: boolean;
-}
-
-export interface ManifestMonitor {
-	readonly monitorId: string;
-	/** Owning agent session id — never the bash_N/watch_N runtime id. */
-	readonly sessionId: string;
-	readonly description: string;
-	readonly runtimeKind: MonitorRuntimeKind;
-	readonly durabilityClass: MonitorDurabilityClass;
-	readonly command?: string;
-	readonly path?: string;
-	readonly event?: "create" | "modify";
-	readonly filter?: string;
-	readonly cwd?: string;
-	readonly approvedParent?: string;
-	readonly createdAt: number;
-	readonly expiresAt: number | null;
-	readonly persistent: boolean;
-	readonly suspended: boolean;
-	readonly lastCheckpoint: TerminalManifestCheckpoint | null;
-	readonly deliveryPaused: boolean;
-	readonly wakeCount: number;
-	readonly fireWindow: { startMs: number; count: number };
-}
-
-export interface ManifestBackgroundSession {
-	readonly id: string;
-	readonly command: string;
-	readonly startedAtMs: number;
-}
-
-export interface TerminalManifest {
-	readonly version: typeof TERMINAL_MANIFEST_VERSION;
-	readonly sessionId: string;
-	readonly monitors: readonly ManifestMonitor[];
-	readonly backgroundSessions: readonly ManifestBackgroundSession[];
-	readonly updatedAt: number;
-}
-
-/** What the monitor tool captures at its call site — the only place the branch inputs live. */
-export interface CommandMonitorSpec {
-	readonly kind: "command";
-	readonly description: string;
-	readonly command: string;
-	readonly filter?: string;
-	readonly cwd?: string;
-	readonly persistent: boolean;
-}
-
-export interface FileMonitorSpec {
-	readonly kind: "file";
-	readonly description: string;
-	readonly path: string;
-	readonly event: "create" | "modify";
-	readonly timeoutMs: number;
-	readonly cwd: string;
-	readonly approvedParent?: string;
-	/** Persistent file watches are the durable `checkpointed-file` class; one-shot ones stay ephemeral. */
-	readonly persistent: boolean;
-}
-export type MonitorSpec = CommandMonitorSpec | FileMonitorSpec;
-
-export interface MonitorRegistration {
-	readonly monitorId: string;
-	readonly spec: MonitorSpec;
-}
+export type {
+	CommandMonitorSpec,
+	FileMonitorSpec,
+	ManifestBackgroundSession,
+	ManifestMonitor,
+	MonitorDurabilityClass,
+	MonitorRegistration,
+	MonitorRuntimeKind,
+	MonitorSpec,
+	TerminalManifest,
+	TerminalManifestCheckpoint,
+	TerminalManifestSession,
+} from "./terminal-manifest-model.ts";
+// The data model is re-exported so every importer keeps this module as the manifest's single
+// entry point; nothing outside had to learn the model module's path.
+export {
+	DURABLE_MONITOR_EXPIRY_MS,
+	TERMINAL_MANIFEST_CHECKPOINT_DEBOUNCE_MS,
+	TERMINAL_MANIFEST_VERSION,
+} from "./terminal-manifest-model.ts";
 
 export function createTerminalManifestStore(session: TerminalManifestSession): SidecarStore<TerminalManifest> {
 	return createSidecarStore({
@@ -282,7 +223,6 @@ export class TerminalManifestWriter {
 			suspended: false,
 			lastCheckpoint: null,
 			deliveryPaused: false,
-			wakeCount: 0,
 			fireWindow: { startMs: createdAt, count: 0 },
 		};
 	}
